refactor: move trends_data module to core
